contract_risk_env/requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
openai
python-dotenv
requests
